% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rwind.r
\name{deg2comp}
\alias{deg2comp}
\title{Wind direction to compass direction}
\usage{
deg2comp(x, bins = c("N", "NNE", "NE", "ENE", "E", "ESE", "SE", "SSE", "S",
  "SSW", "SW", "WSW", "W", "WNW", "NW", "NNW"))
}
\arguments{
\item{x}{A numeric vector of wind directions in degrees.}
}
\value{
Wind directions converted to compass directions
}
\description{
\code{deg2comp} converts horizontal wind speed and direction to uv
components of wind.
}
\examples{
deg2comp(runif(100, 0, 360))
x <- array(runif(1000, 0, 360), dim = c(10, 10, 10))
deg2comp(x)

}
\author{
Ismail SEZEN, \email{sezenismail@gmail.com}
}

