% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wind_functions2.R
\name{ds2uv}
\alias{ds2uv}
\title{Transform direction and speed in U and V components}
\usage{
ds2uv(d, s)
}
\arguments{
\item{u}{direction (degrees).}

\item{v}{speed (m/s).}
}
\value{
"ds2uv" returns a matrix with U and V values
}
\description{
Transform direction and speed in U and V components
}
\note{
Multiple speed and direction values can be procesed.
}
\examples{

\dontrun{

ds2uv(c(32,78,300,234),c(6,2,8,3.2))

}

}
\seealso{
\code{\link{wind.stats}}, \code{\link{wind2raster}}
}
\author{
Javier Fernández-López (jflopez@rjb.csic.es)
}
\keyword{~wind}
