% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rwind.r
\name{hd2uv}
\alias{hd2uv}
\title{Horizontal wind speed & direction to uv}
\usage{
hd2uv(..., deg = T, drop = T)
}
\arguments{
\item{...}{The names of the objects to be converted}

\item{deg}{Should the resulting wind directions are degree or radian?}

\item{drop}{Should be deleted the dimensions of an array which
have only one level?}
}
\value{
uv components of the wind
}
\description{
Convert horizontal wind speed and direction to uv components of wind.
}
\examples{
hd2uv(5, 30)
hd2uv(c(5, 30))
x <- matrix(c(runif(100, -5, 5), round(runif(100, 0, 360), 2)), ncol = 2,
            dimnames = list(NULL, c("h","d")))
hd2uv(x)

}
\author{
Ismail SEZEN, \email{sezenismail@gmail.com}
}

