% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rwind.r
\name{rotax}
\alias{rotax}
\title{Rotate coordinate axis for uv wind components}
\usage{
rotax(..., alfa = 45, deg = T, right = T, drop = T)
}
\arguments{
\item{...}{The names of the objects to be converted.}

\item{alfa}{Rotation angle in degrees or radians.}

\item{deg}{TRUE if alfa is degree.}

\item{right}{TRUE if rotation is in clockwise.}

\item{drop}{Should be deleted the dimensions of an array which have
only one level?}
}
\value{
Rotated uv components
}
\description{
Rotates coordinate axes for uv components of the wind. uv values will be
re-calculated based on new axis. For instance, rotating axis 30 degrees
in clockwise means new and old positive y-directions will have 30 degrees
 between them.
}
\examples{
rotax(1, 1)
rotax(1, 1, alfa = 120)
x <- array(runif(100, -5, 5), dim = c(100, 100, 2))
rotax(x)

}
\author{
Ismail SEZEN, \email{sezenismail@gmail.com}
}

