% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rWind-package.R
\docType{package}
\name{rWind-package}
\alias{rWind-package}
\alias{rWind}
\title{Download, edit and include wind and sea currents data in ecological and evolutionary analysis}
\description{
rWind contain tools for downloading, editing and transforming wind and sea currents data from
Global Forecast System (GFS) and Ocean Surface Current Analyses Real-time (OSCAR). It also allows to use wind and sea currents data to compute the
minimum cost path from wind speed and direction to perform
connectivity analysis.
}
\details{
The complete list of functions can be displayed with \code{library(help =
rWind)}. For more information, please check:
http://allthiswasfield.blogspot.com.es/
}
\seealso{
Useful links:
\itemize{
  \item \url{http://allthiswasfield.blogspot.com.es/}
  \item Report bugs at \url{https://github.com/jabiologo/rWind/issues}
}

}
\author{
Javier Fernández-López

Klaus Schliep

Yurena Arjona

Maintainer: Javier Fernández-López \href{mailto:jflopez.bio@gmail.com}{jflopez.bio@gmail.com}
}
\keyword{package}
