% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rwind.r
\docType{package}
\name{rwind}
\alias{rwind}
\alias{rwind-package}
\title{rwind: A package for processing wind and related data.}
\description{
The rwind package provides four categories of important functions:
\code{\link{deg2comp}}, \code{\link{hd2uv}}, \code{\link{uv2hd}}
and \code{\link{rotax}}.
}
\section{rwind functions}{

The rwind functions ...
}

