% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wind_functions2.R
\name{speed.boundaries}
\alias{speed.boundaries}
\title{Wind speed boundaries}
\usage{
speed.boundaries(wind_series, fun = max)
}
\arguments{
\item{wind_series}{A list of data.frames downloaded by wind.dl function.}

\item{fun}{min or max.}
}
\value{
A numeric vector with max or min speed por each cell in the study
area for this time series.
}
\description{
speed.boundaries computes limits (min and max) of a time series dataset
of wind speeds in the same region. Note that to compute this boundaries for
wind direction makes no sens, so directions are not used.
}
\examples{


# Use wind.dl to download and store wind data of
# the January 3rd 2015 at several hours around New Zealand.

data(wind.series)

# Use wind.stats to compute several measures:

max.speed <- speed.boundaries(wind.series, fun=max)
min.speed<- speed.boundaries(wind.series, fun=min)


}
\references{
https://en.wikipedia.org/wiki/Cross_product
}
\seealso{
\code{\link{wind.dl}}
}
\author{
Javier Fernández-López (jflopez@rjb.csic.es)
}
\keyword{~kwd1}
\keyword{~kwd2}
