% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rwind.r
\name{uv2hd}
\alias{uv2hd}
\title{uv to horizontal speed & direction}
\usage{
uv2hd(..., deg = T, reverse = T, drop = T)
}
\arguments{
\item{...}{The names of the objects to be converted}

\item{deg}{Should the resulting wind directions are degree or radian?}

\item{reverse}{should be added 180 degree (or pi) to the wind directions?}

\item{drop}{Should be deleted the dimensions of an array which
have only one level?}
}
\value{
Speed and direction of the wind
}
\description{
Convert uv components of wind to horizontal wind speed and direction.
}
\examples{
uv2hd(1, 1)
uv2hd(c(1, 1))
x <- array(runif(100, -5, 5), dim = c(100, 100, 2))
uv2hd(x)

}
\author{
Ismail SEZEN, \email{sezenismail@gmail.com}
}

