% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wind_functions2.R
\name{wind.mean}
\alias{wind.mean}
\title{Wind-data mean}
\usage{
wind.mean(x)
}
\arguments{
\item{x}{An object of class \code{rWind_series}}
}
\value{
An object of class \code{rWind}, which is a \code{data.frame}
}
\description{
wind.mean computes the mean (average) of a time series dataset of winds in
the same region. To do this, wind.mean uses U and V vector components of
several wind data.frames stored in a list. Note that, if you want to perform
wind direction and speed average, first you should calculate the mean of U
and V components and then transform it to direction and speed using wind.fit
function from "rWind" package.
}
\note{
For large time series, it could take a while.
}
\examples{


data(wind_series)

#wind_average<- wind.mean(wind_series)


}
\references{
https://en.wikipedia.org/wiki/Cross_product
}
\seealso{
\code{\link{wind.dl}}
}
\author{
Javier Fernández-López (jflopez@rjb.csic.es)
}
\keyword{~kwd1}
\keyword{~kwd2}
