% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rWind-package.R
\docType{data}
\name{wind_data}
\alias{wind_data}
\title{Wind data example}
\format{A data.frame with 652 observations on the following 5 variables:
\describe{ \item{list("time (UTC)")}{a factor with selected time of wind
data } \item{list("latitude (degrees_north)")}{a factor with latitude values
} \item{list("longitude (degrees_east)")}{a factor with longitude values }
\item{list("ugrd10m (m s-1)")}{a factor with U component of wind data }
\item{list("vgrd10m (m s-1)")}{a factor with V component of wind data } }}
\source{
http://allthiswasfield.blogspot.com.es/2016/12/rwind-r-package-released.html
}
\description{
This is an example of wind data obtained with wind.dl function for the
Iberian Peninsula coordenates on 12/February/2015 at 00:00 (UTC)
}
\details{
This data set is the result of:

\code{wind_data<-wind.dl(2015,2,12,0,-10,5,35,45)}
}
\examples{

data(wind_data)
str(wind_data)
head(wind_data)

}
\references{
http://oos.soest.hawaii.edu/erddap/info/NCEP_Global_Best/index.html
}
\keyword{datasets,}
\keyword{download}
\keyword{wind,}
