% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rWind-package.R
\docType{package}
\name{rWind-package}
\alias{rWind-package}
\alias{rWind}
\title{Download, edit and include wind data in ecological and evolutionary analysis}
\description{
rWind contain tools for downloading, editing and transforming wind data from
Global Forecast System (GFS). It also allows to use wind data to compute the
minimum cost path from wind speed and direction to perform
connectivity analysis.
}
\details{
The complete list of functions can be displayed with \code{library(help =
rWind)}. For more information, please check:
http://allthiswasfield.blogspot.com.es/
}
\author{
Javier Fern치ndez-L칩pez

Klaus Schliep

Yurena Arjona

Maintainer: Javier Fern치ndez-L칩pez <jflopez.bio@gmail.com>
}
\keyword{package}
