% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wind_functions2.R
\name{tidy.rWind_series}
\alias{tidy.rWind_series}
\alias{tidy}
\title{Transforming a rWind_series object into a data.frame}
\usage{
\method{tidy}{rWind_series}(x, ...)

tidy(x, ...)
}
\description{
The output of tidy is always a data.frame. It is therefore suited for further
manipulation by packages like dplyr, reshape2, ggplot2 and ggvis.
}
\examples{
data(wind.series)
df <- tidy(wind.series)
}
