% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wind_functions2.R
\name{uv2ds}
\alias{uv2ds}
\title{Transform U and V components in direction and speed}
\usage{
uv2ds(u, v)
}
\arguments{
\item{u}{U component.}

\item{v}{U component.}
}
\value{
"uv2ds" returns a matrix with direction and speed values
}
\description{
Transform U and V components in direction and speed
}
\note{
Multiple U and V values can be procesed.
}
\examples{

\dontrun{

uv2ds(c(1,1,3,1), c(1,1.7,3,1))

}

}
\seealso{
\code{\link{wind.mean}}, \code{\link{wind2raster}}
}
\author{
Javier Fernández-López (jflopez@rjb.csic.es)
}
\keyword{~wind}
