% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wind_functions2.R
\name{wind.fit_int}
\alias{wind.fit_int}
\title{Wind-data fit_int}
\usage{
wind.fit_int(tmpx)
}
\arguments{
\item{X}{downloaded data by wind.dl function from "rWind" package.}
}
\value{
data.frame
}
\description{
wind.fit_int is used internaly by wind.dl to transform downloaded data from
GFS. wind.fit_int applies trigonometry tools to transform U and V vector wind
components in wind direction and speed features. It also transforms 0-360
longitude notation obtained from GFS data into -180/180 longitude notation.
Moreover, it cleans dates names and sorts the data by latitude.
}
\note{
This function is used internaly by wind.dl
}
\examples{

wind.dl(2015,2,12,0,2015,2,12,0,-10,5,35,45)

}
\references{
https://en.wikipedia.org/wiki/Cross_product
}
\seealso{
\code{\link{wind.dl}}, \code{\link{wind.stats}},
\code{\link{wind2raster}}
}
\author{
Javier Fernández-López (jflopez@rjb.csic.es)
}
\keyword{internal}
\keyword{~gfs}
\keyword{~wind}
