% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wind_functions2.R
\name{wind.stats}
\alias{wind.stats}
\title{Wind-data stats}
\usage{
wind.stats(wind_series, fun = mean)
}
\arguments{
\item{wind_series}{A list of data.frames downloaded by wind.dl function.}

\item{fun}{any stats function.}
}
\value{
An object of class "rWind"
}
\description{
wind.stats computes stats (mean, median, etc.) of a time series dataset
of winds in the same region. To do this, wind.stats uses U and V vector
components of several wind data.frames stored in a list.
}
\note{
Date and time are given by the first element of the time series
}
\examples{


# Use wind.dl to download and store wind data of
# the January 3rd 2015 at several hours around New Zealand.

data(wind.series)

# Use wind.stats to compute several measures:

wind.mean<- wind.stats(wind.series, fun=mean)
wind.median<- wind.stats(wind.series, fun=median)


}
\references{
https://en.wikipedia.org/wiki/Cross_product
}
\seealso{
\code{\link{wind.dl}}
}
\author{
Javier Fernández-López (jflopez@rjb.csic.es)
}
\keyword{~kwd1}
\keyword{~kwd2}
