% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wind_functions2.R
\name{wind2raster}
\alias{wind2raster}
\title{Wind-data to raster file}
\usage{
wind2raster(x)
}
\arguments{
\item{x}{an "rWind list" obtained by wind.fit}
}
\value{
A raster stack or a list of raster stacks representing wind direction
and speed.
}
\description{
wind2raster crates a raster stack (gridded) with 2 layers: wind speed and
wind direction from the output of wind.dl function from "rWind" package.
Latitude and logitude values are used to locate raster file and to create
raster using rasterFromXYZ function from raster package.
}
\details{
WGS84 datum (non-projected) CRS is selected by default to build the raster
file.
}
\examples{

# Download wind for Iberian Peninsula region at 2015, February 12, 00:00
# wind_data <- wind.dl(2015,2,12,0,-10,5,35,45)

#data(wind_data)

# Fit downloaded dataset to be plotted
#wind_fitted_data <- wind.fit(wind_data)

# Create raster file from the data.frame created by wind.fit, representing wind speed.
#wind2raster(wind_fitted_data, type="speed")

}
\seealso{
\code{\link{wind.dl}}
}
\author{
Javier Fernández-López (jflopez@rjb.csic.es)
}
\keyword{~gfs}
\keyword{~wind}
