% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wind_functions2.R
\name{wind2raster_int}
\alias{wind2raster_int}
\title{Wind-data to raster file}
\usage{
wind2raster_int(x)
}
\arguments{
\item{x}{an object of class \code{rWind}}
}
\value{
A raster file representing wind direction, wind speed or both of the
study area.
}
\description{
wind2raster_int crates a raster file (gridded) from a data.frame created by
wind.fit function from "rWind" package. Latitude and logitude values are used
to locate raster file and create raster resolution using rasterFromXYZ
function from raster package. As raster files only can store one field of
information, you should choose between direction (by default, type="dir")
and speed (type="speed") to be represented by the new raster file.
}
\details{
WGS84 datum (non-projected) CRS is selected by default to build the raster
file.
}
\seealso{
\code{\link{wind.dl}}, \code{\link{wind2raster}}
}
\author{
Javier Fernández-López (jflopez@rjb.csic.es)
}
\keyword{internal}
\keyword{~gfs}
\keyword{~wind}
